feat(workers): knowledge_search auto-renders top result inline
knowledge_search now ships the chart widget for the top result as part
of its own tools/call response, so ChatGPT renders the chart inline
without needing to chain into open_chart_ui. Fixes the "asked twice"
pattern where the model would reply with a link instead of the card.

  - knowledge_search: new appUiResource / extraMeta /
    extraContentBlocks hooks. Handler lifts pub_id, embed_url,
    image_url, dark_mode, width, height to the output root when
    results[0].card_id exists; widget reads the same keys it already
    reads on open_chart_ui.

  - _chart_widget.ts: shared module holding WIDGET_HTML, the baked
    template variant, PNG fetch helpers, URL builders, and the
    AppUiResource factory. Both tools register the same
    ui://tako/embed/chart URI and bundle.

  - mcp.ts: dedupe registerResource() / ResourceTemplate
    registrations across tools that share a widget URI (the SDK
    throws on duplicates). Thread output to dynamic
    resolveUriFromInput so knowledge_search can derive the per-call
    URI from results[0].card_id.

  - Descriptions rewritten: knowledge_search drops the "MUST chain
    into open_chart_ui" paragraph and explains the auto-render plus
    multi-card narration ("when results.length > 1, mention the
    other titles..."). open_chart_ui reframed as the explicit
    re-render tool for create_chart chains, non-top results, and
    pub_id references from prior turns.
