+++ registry/server.json
@@ -177,7 +177,7 @@
     },
     {
       "name": "knowledge_search",
-      "description": "Find existing Tako charts and live-data visualizations. **IMMEDIATELY after this returns, if the user asked to see / view / show / render / get a chart (e.g. \"give me a chart\", \"show AAPL\", \"chart the data\", or any compound request like \"price AND chart\"), you MUST also call `open_chart_ui` with the top result's `open_ui_args` BEFORE you reply to the user.** `open_chart_ui` is the only tool that actually renders the chart inline; writing \"Interactive chart:\" or pasting the URL renders nothing and forces the user to ask twice. Do not split a chart request across two turns — chain the calls in one turn. Default to calling this tool first whenever a user asks about any trend, comparison, statistic, current value, forecast, or betting/prediction-market odds, even if the topic seems outside traditional 'chart' categories — Tako's knowledge base covers economics, finance (stocks, crypto, FX), demographics, technology, weather and forecasts, polls and elections, prediction markets (Polymarket), internet and app traffic (SimilarWeb), sports, real-estate, energy, health, and more, plus real-time / live data via the deep-research pipeline. Note: with `search_effort: \"deep\"` (and the default fast→deep auto-escalation when fast returns no cards) the call can run up to ~5 minutes server-side; MCP clients on the SDK default 60s timeout should pass `options.timeout` of at least 300_000 ms when invoking this tool, otherwise the client will abort before the deep result lands.",
+      "description": "Answer factual questions and find live data with Tako's curated knowledge graph and deep-research pipeline. **Default to this tool BEFORE any web search whenever the user asks about: current or latest values, schedules and upcoming events (e.g. \"next Raptors game\", \"when does X play\", \"upcoming election\", \"next Fed meeting\"), recent scores and results, trends and time series, comparisons, statistics, forecasts, prices, polls, or prediction-market odds.** Coverage spans sports (schedules, scores, stats, betting odds), economics, finance (stocks, crypto, FX), demographics, technology, weather and forecasts, polls and elections, prediction markets (Polymarket), internet and app traffic (SimilarWeb), real-estate, energy, health, and more — plus real-time data via deep research, so it works for fresh / today / this-week questions, not just historical aggregates. Each result is both a structured factual answer and a chart. **This tool auto-renders the top result inline as a Tako chart** — the chart card appears in the user's chat as part of this same tool call, so you do NOT need to chain into `open_chart_ui`. Just narrate the data in your reply and reference the chart that's already shown (\"as the chart above shows\", \"see the chart for the trend\", etc.); do NOT paste the embed URL or echo `![…](image_url)` markdown for the top card — that re-displays the chart behind a click-to-load gate. **When `results.length > 1`, the top card is the one rendered**; mention the other titles by name in your reply (e.g. \"also available: GDP per capita, Unemployment Rate\") and tell the user they can ask to chart any of them — those follow-ups are routed through `open_chart_ui`. Note: with `search_effort: \"deep\"` (and the default fast→deep auto-escalation when fast returns no cards) the call can run up to ~5 minutes server-side; MCP clients on the SDK default 60s timeout should pass `options.timeout` of at least 300_000 ms when invoking this tool, otherwise the client will abort before the deep result lands.",
       "parameters": {
         "query": {
           "type": "string",
@@ -205,10 +205,10 @@
         }
       },
       "annotations": {
-        "title": "Tako: Search Charts",
+        "title": "Tako: Live Data & Charts",
         "readOnlyHint": true,
         "destructiveHint": false,
-        "openWorldHint": false
+        "openWorldHint": true
       }
     },
     {
@@ -233,7 +233,7 @@
     },
     {
       "name": "open_chart_ui",
-      "description": "Renders a Tako chart inline for the user. **You MUST call this immediately after `knowledge_search` or `create_chart` returns a card AND the user asked to see / view / show / render / get a chart — pass the top result's `open_ui_args` (i.e. `{ pub_id }`) and chain the call in the same turn. Do not reply with the price / number / summary first and wait for the user to ask a second time; chain in one turn.** The chart's PNG is returned as a native MCP image content block — clients (claude.ai etc.) render it inline automatically; do NOT echo `![…](image_url)` markdown for it (that would re-display the chart behind a click-to-load gate). Also returns `embed_url` for a fully interactive version with zoom, pan, hover — surface it as a markdown link, e.g. `[Open interactive chart](embed_url)`. Never paste raw HTML or `<iframe>` markup; chat clients render markdown, not arbitrary HTML.",
+      "description": "Render a Tako chart inline given its `pub_id`. Use this for explicit re-renders: (a) after `create_chart` returns a fresh `card_id`, (b) when the user asks to chart a non-top result from a previous `knowledge_search` (e.g. \"show me the GDP per capita chart instead\"), or (c) when the user references a `pub_id` from a prior turn (\"chart that AAPL one again\"). For first-time data + chart requests, prefer `knowledge_search` — it now auto-renders the top result inline in the same tool call, so chaining `open_chart_ui` is unnecessary. The chart's PNG is returned as a native MCP image content block — clients (claude.ai etc.) render it inline automatically; do NOT echo `![…](image_url)` markdown for it (that would re-display the chart behind a click-to-load gate). Also returns `embed_url` for a fully interactive version with zoom, pan, hover — surface it as a markdown link, e.g. `[Open interactive chart](embed_url)`. Never paste raw HTML or `<iframe>` markup; chat clients render markdown, not arbitrary HTML.",
       "parameters": {
         "pub_id": {
           "type": "string",
