+++ workers/src/index.test.ts
@@ -135,27 +135,28 @@ describe("worker routing", () => {
       "wait_for_report",
     ]);
 
-    // MCP Apps: `open_chart_ui` ships a widget bundle, so its tool listing
-    // must carry the widget URI under BOTH `_meta.ui.resourceUri` (the
-    // open MCP Apps spec, read by claude.ai / VS Code / Goose) AND
-    // `_meta["openai/outputTemplate"]` (the OpenAI Apps SDK namespace,
-    // read by ChatGPT — without it ChatGPT loads the widget but never
-    // pipes structuredContent into `window.openai.toolOutput`, so the
-    // widget stays on its loading state forever). Other tools ship no
-    // widget and should declare neither field.
-    const openChart = body.result.tools.find((t) => t.name === "open_chart_ui");
-    // Three metadata keys, all carrying the same widget URI — matches what
-    // OpenAI's official `@modelcontextprotocol/ext-apps` helper emits.
-    // Different hosts read different keys (and different versions of the
-    // same host may have read different keys historically), so we set
-    // them all and let each host pick what it expects.
-    expect(openChart?._meta).toMatchObject({
-      ui: { resourceUri: "ui://tako/embed/chart" },
-      "ui/resourceUri": "ui://tako/embed/chart",
-      "openai/outputTemplate": "ui://tako/embed/chart",
-    });
+    // MCP Apps: BOTH `open_chart_ui` AND `knowledge_search` ship the
+    // chart widget bundle (knowledge_search auto-renders the top
+    // result inline so the model doesn't have to chain into
+    // open_chart_ui). Both tools' listings must carry the widget URI
+    // under all three metadata keys: `_meta.ui.resourceUri` (open MCP
+    // Apps spec, read by claude.ai / VS Code / Goose), the legacy
+    // flat `_meta["ui/resourceUri"]` (older host readers), and
+    // `_meta["openai/outputTemplate"]` (ChatGPT's Apps SDK — without
+    // it the widget loads but `window.openai.toolOutput` never
+    // populates). Other tools ship no widget and should declare
+    // none of these fields.
+    const widgetTools = new Set(["open_chart_ui", "knowledge_search"]);
+    for (const name of widgetTools) {
+      const tool = body.result.tools.find((t) => t.name === name);
+      expect(tool?._meta).toMatchObject({
+        ui: { resourceUri: "ui://tako/embed/chart" },
+        "ui/resourceUri": "ui://tako/embed/chart",
+        "openai/outputTemplate": "ui://tako/embed/chart",
+      });
+    }
     for (const t of body.result.tools) {
-      if (t.name === "open_chart_ui") continue;
+      if (widgetTools.has(t.name)) continue;
       const meta = t._meta as
         | {
             ui?: unknown;
